Allow start iteration to be float.

+++ tools/reblock.py
@@ -30,7 +30,7 @@ def parse_args(args):
     """
 
     parser = argparse.ArgumentParser(description = __doc__)
-    parser.add_argument('-s', '--start', type=int, dest='start_time',
+    parser.add_argument('-s', '--start', type=float, dest='start_time',
                         default=0, help='Imaginary time after which we '
                         'gather statistics.  Default: 0')
     parser.add_argument('-l', '--multi-sim', action='store_true',
